requirements.txt
# base
drf-yasg~=1.20.0
django~=3.2.7
djangorestframework~=3.12.4
gunicorn~=20.1.0
selenium~=4.1.0

# database
psycopg2~=2.9.3